requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
pillow==10.4.0
pycparser==2.22
pycryptodomex==3.21.0
pyOpenSSL==24.2.1
python-alipay-sdk==3.3.0
sqlparse==0.5.1
tzdata==2024.2
